feat: move dashboard common components to /dashboard/common folder
