+++ src/app/dashboard/club/announcement/detail/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { cn } from "@/lib/utils";
-import BackButton from "@/components/dashboard/club/common/BackButton";
+import BackButton from "@/components/dashboard/common/BackButton";
 
 const announcement = [
   {
