+++ src/app/dashboard/club/reservation/item/[id]/page.tsx
@@ -1,4 +1,4 @@
-import BackButton from "@/components/dashboard/club/common/BackButton";
+import BackButton from "@/components/dashboard/common/BackButton";
 import InquiryShortcut from "@/components/dashboard/club/reservation/molecules/InquiryShortcut";
 import ItemDetailOverview from "@/components/dashboard/club/reservation/molecules/ItemOverview";
 import ItemDetail from "@/components/dashboard/club/reservation/organisms/ItemDetail";
