+++ src/app/dashboard/club/reservation/item/[id]/purchase/page.tsx
@@ -1,4 +1,4 @@
-import BackButton from "@/components/dashboard/club/common/BackButton";
+import BackButton from "@/components/dashboard/common/BackButton";
 import PurchaseDetail from "@/components/dashboard/club/reservation/organisms/PurchaseDetail";
 import PurchasePanel from "@/components/dashboard/club/reservation/organisms/PurchasePanel";
 import PurchaseSuccessModal from "@/components/dashboard/club/reservation/organisms/PurchaseSuccessModal";
