+++ src/components/dashboard/club/reservation/molecules/ReservationSearch.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import Chip from "@/components/common/Chip";
-import SearchBar from "@/components/dashboard/club/common/SearchBar";
+import SearchBar from "@/components/dashboard/common/SearchBar";
 
 const filterList = [
   { name: "전체 보기", value: "all" },
