+++ src/components/dashboard/club/reservation/organisms/InquiryList.tsx
@@ -3,8 +3,8 @@
 import { useState } from "react";
 import DateFilter, {
   type DateRange,
-} from "@/components/dashboard/club/common/DateFilter";
-import Pagination from "@/components/dashboard/club/common/Pagination";
+} from "@/components/dashboard/common/DateFilter";
+import Pagination from "@/components/dashboard/common/Pagination";
 import InquiryTable from "@/components/dashboard/club/reservation/molecules/InquiryTable";
 
 export default function InquiryList() {
